requirements.txt
fastapi
uvicorn
google-genai
google-generativeai
python-dotenv
pydantic
